tests.sh: Add background ping while bbperf is running
We want to measure the latency observed by other traffic
than the bbperf bulk UDP workload.

The qdisc should penalize the bbperf UDP flow that is 
trying to consume all the bandwidth on the link (in search
of bufferbloat on the link).

This is a simple approach using the ping command.
It waits 5 seconds before starting the ping, in-order to
give bbperf some time to rampup bandwidth usage.
Observe if the ping packets see any packet loss, as the
goal is to avoid any packet loss for other traffic than
the bbperf bulk UDP workload.

Signed-off-by: Jesper Dangaard Brouer <jesper@cloudflare.com>