setup.sh: disable veth multiqueue setup per default
This is to make sure our ping packets travels in the same queue
as the bbperf UDP traffic.  With multiqueue the ping packets
might not get affected by the UDP packets (if it ends up travling
on a different queue).  Our goal is to measure if ping packets
gets influenced by the bbperf UDP traffic.

Signed-off-by: Jesper Dangaard Brouer <jesper@cloudflare.com>